backend/requirements.txt
Flask==2.1.2
kafka-python==2.0.2
werkzeug==2.0.0

